Make the refactor work.
The files in the `lib` directory all extend the base `lib/mkvb.rb` class,
and each focuses on one specific aspect of the mkvm process: building
the Kickstart line, creating the ISO, working with VMware.

Plugins are fully supported, and each plugin has numerous opportunities
to interact with the process. Two default plugins are provided to
perform some basic sanity checking on the IP information for the VM
being created.