Fix network connection settings bug.
The fix is simple, discovering the required data proved otherwise.  In
lieu of spending more hours attempting something with the VMWare API
that may not be possible I created 2 required data structures that map
known entities to their network-specific values.

If we go with this route every user of mkvm.rb will need to add this
structure to their ~/.mkvm.yaml.  The structure is presented in the
README.md file.  Specific values for CMM will be provided if we decide
to merge.