Parameterize more options; IPAM add/remove URIs
This change removes the hard-coded @covermymeds.com addresses, and
instead adds parameters for SMTP server, from and to email addresses.
These values can be defaulted in `~/.mkvm.yaml`, as well as defined on
the command line.

Only if all three values are set will `rmvm.rb` send an email.

This change also better abstracts the IPAM add and remove URLs.  The
`--auto-uri` parameter in `mkvm.rb` has been renamed to `--add-uri` to
reflect that this is the URI that will add a new IP address.  The
`--auto-uri` parameter in `rmvm.rb` has been renamed to `--del-uri` to
reflect that this is the URI that will remove an IP address.

The CMM-specific URI components have been removed from the Ruby scripts.
Instead, the URIs should be specified in the `~/.mkvm.yaml`, and now the
Ruby scripts will perform string substitutions on the supplied value.

Right now, the following tokens will be replaced in the URI strings:

* HOSTNAME => the hostname as specified on the command line
* SUBNET => the subnet as specified on the command line
* USER => the user as specified on the command line

**NOTE**: right now, `rmvm.rb` only replaces `HOSTNAME`.