Make app_env an optional part of the kickstart
We have updated our build process to encapsulate the server app_env into
the hostname, and a custom fact parses the hostname to populate the
app_env fact.

This change removes the app_env value from the Kickstart boot line if
the `--app-env` option is not supplied.  This makes it possible to still
define an app_env if necessary.